RFFT-demo.py, longtermspectrum-aeksjiayan-5s.png
RFFT-demo.py is a demo of estimating the real component of the FFT, for the Long-Term spectrum (LTS) and for the Amplitude Envelope Modulation Spectrum (AEMS). The PNG file longtermspectrum-aeksjiayan-5s.png shows the kind of output provided. AEM is extracted with a peak maximum technique over a moving window, rather than Hilbert Transform.